Update README.md
Main Servlet-class created and implemented
Added Validator class with methods for checking the entered data, classes for creating / connecting to pool of connections.
Added IDao, IService implementation for Employee and Department

The functional interface has been changed to work correctly with it
Changed AbstractDecoratorDao modifiers methods so that methods use inherited classes

Added packaging in pom for web project. Added new queries for working with the database and method in DAO and Service.
Added javadoc.